Added docs, moved port to int datatype instead of short.

+++ src/main/java/utwente/ns/tcp/RTP4Socket.java
@@ -410,7 +410,7 @@ public class RTP4Socket implements IReceiveListener, Closeable {
         return state == SocketState.CLOSED;
     }
 
-    public short getPort(){
+    public int getPort(){
         return ipSocket.getDstPort();
     }
